Duplicate row from TimelineEditor
